# postman/translate-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Translate API",
    "version": "0.3.0",
    "description": "**Translate API** — production-grade translation, i18n batch and multilingual dictionary lookup, powered by **Qwen3-14B-AWQ** served on vLLM.\n\n## Highlights\n- `POST /v1/translate` — single string or batch (up to 100 items, each up to 50 000 chars).\n- `POST /v1/json` — i18n array of strings, order-preserving.\n- `POST /v1/translate-html` — translate HTML preserving tag structure, with inline-aware placeholders and a brand-name `ignore_terms` whitelist. Auto-fixes minor parse issues; rejects severely malformed HTML with **422**.\n- `POST /v1/dict` — rich dictionary entry for language learners (phonetic, definitions, examples, phrases, synonyms/antonyms, mnemonic tips).\n- Auto language detection with per-item `detected_source_lang` when `source_lang=auto`; plain `list[str]` response when an explicit source is given.\n- Script-purity post-validation with single retry to catch mixed-script outputs.\n\n## Quick start\n1. Subscribe on RapidAPI to obtain your `X-RapidAPI-Key`.\n2. `POST /v1/translate` with `{ \"text\": \"Hello\", \"target_lang\": \"vi\" }`.\n3. Read `translations[*]` — strings or `{translated_text, detected_source_lang}` objects depending on `source_lang`.",
    "termsOfService": "https://translate.spacecloud.fit/terms",
    "contact": {
      "name": "Translate API Support",
      "url": "https://translate.spacecloud.fit",
...
    },
    "license": {
...
    },
...
  "servers": [
...
  "paths": {
    "/v1/translate": {
      "post": {
        "tags": [
          "translate"
        ],
        "summary": "Translate single text or a batch of strings.",
        "description": "Translate one string or up to 100 strings in one request. The response shape depends on `source_lang`:\n- `source_lang=auto` → `translations` is an array of `{translated_text, detected_source_lang}`.\n- explicit `source_lang` → `translations` is an array of strings, in the same order as the input.\n\nWhen `text` is a single string, `translations` is still an array (length 1) for shape consistency.",
        "operationId": "translate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateRequest"
              },
              "examples": {
                "single_auto": {
                  "summary": "Single text + auto detect",
                  "value": {
                    "text": "Hello world",
                    "target_lang": "vi"
                  }
                },
                "single_explicit": {
                  "summary": "Single text + explicit source",
                  "value": {
                    "text": "Tôi yêu Việt Nam",
                    "source_lang": "vi",
                    "target_lang": "en"
                  }
                },
                "batch_explicit": {
                  "summary": "Batch + explicit source",
                  "value": {
                    "text": [
                      "Good morning",
                      "Thank you"
                    ],
                    "source_lang": "en",
                    "target_lang": "vi"
                  }
                },
                "batch_auto_mixed": {
                  "summary": "Batch + auto on mixed-language input",
                  "value": {
                    "text": [
                      "こんにちは",
                      "안녕하세요"
                    ],
                    "target_lang": "vi"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslateResponse"
                },
                "examples": {
                  "auto_with_detected": {
                    "summary": "Auto: each item carries detected_source_lang",
                    "value": {
                      "request_id": "03f0ece3-9ad8-4b52-ad0c-40da09487722",
                      "processing_time_ms": 126,
                      "translations": [
                        {
                          "translated_text": "Xin chào thế giới",
                          "detected_source_lang": "en"
                        }
                      ]
                    }
                  },
                  "explicit_strings": {
                    "summary": "Explicit source: list of strings",
                    "value": {
                      "request_id": "f4bfc370-747e-40c7-a3c0-eb1cb86683af",
                      "processing_time_ms": 86,
                      "translations": [
                        "Chào buổi sáng",
                        "Cảm ơn"
                      ]
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error (invalid lang code, empty text, exceeds size limits, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream LLM error (vLLM backend unreachable or returned error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/json": {
      "post": {
        "tags": [
          "translate"
        ],
        "summary": "Translate an array of strings (i18n batch).",
        "description": "Same response shape as `/v1/translate`. Use this endpoint when your input is always an array (e.g. localization keys for a UI). Order is preserved.",
        "operationId": "translateJson",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonTranslateRequest"
              },
              "examples": {
                "i18n_explicit": {
                  "summary": "i18n keys (explicit en→ja)",
                  "value": {
                    "texts": [
                      "Welcome",
                      "Sign in",
                      "Sign up"
                    ],
                    "source_lang": "en",
                    "target_lang": "ja"
                  }
                },
                "i18n_auto": {
                  "summary": "Auto detect on mixed batch",
                  "value": {
                    "texts": [
                      "こんにちは",
                      "안녕하세요"
                    ],
                    "target_lang": "vi"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translations in the same order as `texts`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonTranslateResponse"
                },
                "examples": {
                  "auto_with_detected": {
                    "summary": "Auto output",
                    "value": {
                      "request_id": "263aafcd-ea31-423e-bd4b-1591d2f6cbe3",
                      "processing_time_ms": 109,
                      "translations": [
                        {
                          "translated_text": "Xin chào",
                          "detected_source_lang": "ja"
                        },
                        {
                          "translated_text": "Xin chào",
                          "detected_source_lang": "ko"
                        }
                      ]
                    }
                  },
                  "explicit_strings": {
                    "summary": "Explicit output",
                    "value": {
                      "request_id": "8a24186d-df3b-4843-8caf-63702cc1e84c",
                      "processing_time_ms": 71,
                      "translations": [
                        "ようこそ",
                        "サインイン",
                        "サインアップ"
                      ]
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error (invalid lang code, empty text, exceeds size limits, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream LLM error (vLLM backend unreachable or returned error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/translate-html": {
      "post": {
        "tags": [
          "translate"
        ],
        "summary": "Translate HTML preserving structure.",
        "description": "Translate HTML while preserving tags, attributes, structure and inline formatting. Pipeline:\n1. Parse with `lxml` (recover mode) → fallback to `html5lib` if errors are heavy.\n2. Score health: `clean` / `minor` / `moderate` / `severe`. `severe` returns **422** asking the caller to fix the HTML first.\n3. Walk DOM. Skip `<script>`, `<style>`, `<noscript>`, `<svg>`, `<math>`, `<template>`, `<textarea>`, `<pre>`, `<code>`, `<kbd>`, `<samp>`, `<var>`.\n4. Build inline-aware segments at leaf-block level using XLIFF-style `<g id=N>...</g>` placeholders so inline `<b>`, `<i>`, `<a>`, etc. survive any reordering by the model.\n5. `ignore_terms` — replace exact word-boundary matches with `<x-keep id=N/>` placeholders; restored verbatim after translation.\n6. Batch translate via vLLM continuous batching (concurrency=16).\n7. Re-insert into DOM, serialize, re-parse and verify tag count.\n\n**Translatable attributes** (when `translate_attributes=true`): `alt`, `title`, `aria-label`, `placeholder`, `<meta name=description|keywords>` content.",
        "operationId": "translateHtml",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateHtmlRequest"
              },
              "examples": {
                "simple_inline": {
                  "summary": "Simple HTML with inline tags",
                  "value": {
                    "html": "<p>Hello <b>red</b> car!</p>",
                    "target_lang": "vi"
                  }
                },
                "ignore_terms_brand": {
                  "summary": "Brand names preserved via ignore_terms",
                  "value": {
                    "html": "<p>Apple makes the new MacBook Pro and iPhone. Buy now!</p>",
                    "source_lang": "en",
                    "target_lang": "vi",
                    "ignore_terms": [
                      "Apple",
                      "MacBook Pro",
                      "iPhone"
                    ]
                  }
                },
                "attributes": {
                  "summary": "Translate alt + title attributes",
                  "value": {
                    "html": "<img src=\"x.jpg\" alt=\"A picture of a cat\" title=\"Click to enlarge\"><p>Image gallery</p>",
                    "source_lang": "en",
                    "target_lang": "vi"
                  }
                },
                "skip_script_code": {
                  "summary": "Skip script/style/code blocks",
                  "value": {
                    "html": "<p>Use <code>print()</code> to debug.</p><script>console.log(\"skip\")</script>",
                    "source_lang": "en",
                    "target_lang": "vi"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translated HTML (clean/minor/moderate health).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslateHtmlResponse"
                },
                "examples": {
                  "simple_inline": {
                    "summary": "Inline tag preserved",
                    "value": {
                      "request_id": "e3a1109d-4dda-4c4e-a207-8e7a4efe7d85",
                      "processing_time_ms": 192,
                      "html": "<p>Chào xe <b>đỏ</b>!</p>",
                      "detected_source_lang": "en",
                      "health": {
                        "health": "clean",
                        "error_rate": 0.0,
                        "structure_diff": 0.0,
                        "errors_total": 0,
                        "fatals_total": 0,
                        "parse_tier": "fragment_wrap"
                      },
                      "segments_translated": 1,
                      "chars_translated": 28,
                      "warnings": []
                    }
                  },
                  "ignore_terms": {
                    "summary": "Brand names kept",
                    "value": {
                      "request_id": "e01239ca-ac36-4cea-a9f7-628b046f3e2b",
                      "processing_time_ms": 113,
                      "html": "<p>Apple tạo ra MacBook Pro mới và iPhone. Mua ngay!</p>",
                      "detected_source_lang": "en",
                      "health": {
                        "health": "clean",
                        "error_rate": 0.0,
                        "structure_diff": 0.0,
                        "errors_total": 0,
                        "fatals_total": 0,
                        "parse_tier": "fragment_wrap"
                      },
                      "segments_translated": 1,
                      "chars_translated": 56,
                      "warnings": []
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "HTML too malformed (`severe` health). Caller must fix HTML before retry. Either standard validation error or `html_too_malformed` payload.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/HtmlTooMalformedError"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Upstream LLM error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TranslateRequest": {
        "type": "object",
        "required": [
          "text",
          "target_lang"
        ],
        "properties": {
          "text": {
            "description": "Text to translate. Either a single string or an array (max 100 items, each item up to 50 000 characters).",
            "oneOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 50000
              },
              {
                "type": "array",
                "minItems": 1,
                "maxItems": 100,
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 50000
                }
              }
            ],
            "example": [
              "Good morning",
              "Thank you"
            ]
          },
          "source_lang": {
            "type": "string",
            "default": "auto",
            "pattern": "^(auto|[a-z]{2,3}(-[A-Z]{2})?)$",
            "description": "`auto` lets the model detect; otherwise an ISO 639-1 code (`vi`, `en`, `ja`, `zh`, `ko`, `fr`, `de`, ...).",
            "example": "en"
          },
          "target_lang": {
            "type": "string",
            "minLength": 2,
            "maxLength": 8,
            "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
            "description": "Required ISO 639-1 target language. Cannot be `auto`.",
            "example": "vi"
          }
        }
      },
      "TranslationDetected": {
        "type": "object",
        "description": "Returned when `source_lang=auto`: each item carries its own detected language.",
        "required": [
          "translated_text",
          "detected_source_lang"
        ],
        "properties": {
          "translated_text": {
            "type": "string",
            "description": "Translated text."
          },
          "detected_source_lang": {
            "type": "string",
            "description": "ISO 639-1 code detected from the source text.",
            "example": "en"
          }
        }
      },
      "TranslateResponse": {
        "type": "object",
        "required": [
          "request_id",
          "processing_time_ms",
          "translations"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "processing_time_ms": {
            "type": "integer",
            "minimum": 0
          },
          "translations": {
            "description": "Translations in the same order as the input. When `source_lang=auto`, each item is `{translated_text, detected_source_lang}`. When an explicit `source_lang` is given, items are plain strings.",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TranslationDetected"
                }
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          }
        }
      },
      "JsonTranslateRequest": {
        "type": "object",
        "required": [
          "texts",
          "target_lang"
        ],
        "properties": {
          "texts": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50000
            },
            "description": "Array of strings to translate (max 100 items, each up to 50 000 chars).",
            "example": [
              "Welcome",
              "Sign in",
              "Sign up"
            ]
          },
          "source_lang": {
            "type": "string",
            "default": "auto",
            "pattern": "^(auto|[a-z]{2,3}(-[A-Z]{2})?)$",
            "example": "en"
          },
          "target_lang": {
            "type": "string",
            "minLength": 2,
            "maxLength": 8,
            "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
            "description": "Required ISO 639-1 target language.",
            "example": "ja"
          }
        }
      },
      "JsonTranslateResponse": {
        "type": "object",
        "required": [
          "request_id",
          "processing_time_ms",
          "translations"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "processing_time_ms": {
            "type": "integer",
            "minimum": 0
          },
          "translations": {
            "description": "Translations in the same order as the input. When `source_lang=auto`, each item is `{translated_text, detected_source_lang}`. When an explicit `source_lang` is given, items are plain strings.",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TranslationDetected"
                }
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          }
        }
      },
      "TranslateHtmlRequest": {
        "type": "object",
        "required": [
          "html",
          "target_lang"
        ],
        "properties": {
          "html": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000000,
            "description": "HTML to translate (max 5 000 000 chars).",
            "example": "<p>Hello <b>world</b>!</p>"
          },
          "source_lang": {
            "type": "string",
            "default": "auto",
            "pattern": "^(auto|[a-z]{2,3}(-[A-Z]{2})?)$",
            "example": "en"
          },
          "target_lang": {
            "type": "string",
            "minLength": 2,
            "maxLength": 8,
            "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
            "example": "vi"
          },
          "translate_attributes": {
            "type": "boolean",
            "default": true,
            "description": "Translate `alt`, `title`, `aria-label`, `placeholder` attributes and `<meta name=\"description\">` content. Set false for text-only translation."
          },
          "ignore_terms": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "maxItems": 500,
            "default": [],
            "description": "Words or multi-word phrases to keep verbatim (brand names, product names, technical jargon). Word-boundary match. Default case-sensitive.",
            "example": [
              "Apple",
              "MacBook Pro",
              "iPhone 15 Pro Max"
            ]
          },
          "ignore_case": {
            "type": "boolean",
            "default": false,
            "description": "Case-insensitive match for `ignore_terms`. Default false (`Apple` ≠ `apple`)."
          }
        }
      },
...
      "TranslateHtmlResponse": {
        "type": "object",
        "required": [
          "request_id",
          "processing_time_ms",
          "html",
          "health",
          "segments_translated",
          "chars_translated",
          "warnings"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "processing_time_ms": {
            "type": "integer",
            "minimum": 0
          },
          "html": {
            "type": "string",
            "description": "Translated HTML, structure preserved."
          },
          "detected_source_lang": {
            "type": "string",
            "nullable": true,
            "description": "Dominant detected language when `source_lang=auto`. Null for explicit lang.",
            "example": "en"
          },
          "health": {
            "$ref": "#/components/schemas/HtmlHealth"
          },
          "segments_translated": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of text/attribute segments translated."
          },
          "chars_translated": {
            "type": "integer",
            "minimum": 0,
            "description": "Total characters sent to the model."
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Soft warnings: parser auto-fixes, lost ignore-term placeholders, post-translate tag count drift."
          }
        }
      },
      "HtmlTooMalformedError": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "object",
            "required": [
              "error",
              "health",
              "metrics"
            ],
            "properties": {
              "error": {
                "type": "string",
                "enum": [
                  "html_too_malformed"
                ]
              },
              "health": {
                "type": "string",
                "enum": [
                  "severe"
                ]
              },
              "metrics": {
                "type": "object",
                "properties": {
                  "error_rate": {
                    "type": "number"
                  },
                  "structure_diff": {
                    "type": "number"
                  },
                  "errors_total": {
                    "type": "integer"
                  },
                  "fatals_total": {
                    "type": "integer"
                  }
                }
              },
              "errors_sample": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "line": {
                      "type": "integer"
                    },
                    "column": {
                      "type": "integer"
                    },
                    "severity": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "fatal_markers": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "suggestion": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            ],
            "description": "Error detail (string for upstream errors, array of validation errors from FastAPI)."
          }
        }
      }
    }
  }
}